requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.2
django-bootstrap-form==3.4
django-bootstrap4==2.3.0
djangorestframework==3.12.1
Pillow==7.2.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
